migrate account and errors packages to work as packages
under pkg/ instead of runnable commands (under cmd/)
